Rename field reference utility for clearer semantics
